+++ docs/tr1/CHANGELOG.md
@@ -2,7 +2,7 @@
 - reworked screen resolutions
     - removed "screen resolution" option
     - added aspect ratio mode (Graphic options → Rendering → Aspect mode)
-    - added sizer option (Graphic options → Rendering → Sizer)
+    - added window border option (Graphic options → Rendering → Borders)
     - added integer upscaling option (Graphic options → Rendering → Upscaling factor)
     - renamed "fbo filter" option to "upscaling filter" (Graphic options → Rendering → Upscaling filter)
 - added French translation